move standard config to separate package
