+++ validation/standard/standard-config-mainnet.toml
@@ -22,6 +22,9 @@ decimals = [6, 6]
 blob_base_fee_scalar = [0, 10_000_000]
 base_fee_scalar = [0, 10_000_000]
 
+[system_config]
+gas_limit = [1000, 60_000_000]
+
 [not-yet-covered]
 batcher_inbox_address = "0xff000...000{chainId}"                  # this is only a convention and can't be checked on chain
 batcher_hash = "bytes32(uint256(uint160(batchSubmitterAddress)))" # this can be read from SystemConfig, but set arbitrarily
