+++ validation/standard/standard-config-sepolia.toml
@@ -21,3 +21,6 @@ scalar = [684_000, 1_000_000]
 decimals = [6, 6]
 blob_base_fee_scalar = [0, 10_000_000]
 base_fee_scalar = [0, 10_000_000]
+
+[system_config]
+gas_limit = [1000, 60_000_000]
